requirements.txt
asgiref==3.3.4
beautifulsoup4==4.9.3
bootstrap4==0.1.0
dj-database-url==0.5.0
Django==3.2
django-bootstrap-static==4.2.1
django-bootstrap4==2.3.1
django-crispy-forms==1.11.2
django-formtools==2.3
django-heroku==0.3.1
fprint==1.0
gunicorn==20.1.0
psycopg2==2.8.6
psycopg2-binary==2.8.6
pytz==2021.1
soupsieve==2.2.1
sqlparse==0.4.1
whitenoise==5.2.0
